pyproject.toml
[project]
name = "sign-language-segmentation"
description = "Sign language pose segmentation model on both the sentence and sign level"
version = "0.0.1"
authors = [
    { name = "Amit Moryossef", email = "amitmoryossef@gmail.com" },
    { name = "Zifan Jiang", email = "jiang@cl.uzh.ch" }
]
readme = "README.md"
dependencies = [
    "pose-format>=0.8.1",
    "numpy",
    "pympi-ling", # Working with ELAN files in CLI
    "torch",
    "pose_anonymization @ git+https://github.com/sign-language-processing/pose-anonymization", # Used for normalization
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pylint",
    "pytorch-lightning",
    "wandb",
    "matplotlib",
    "scikit-learn",
    "psycopg2-binary" # to fetch the annotations from the database
]

[tool.yapf]
based_on_style = "google"
column_limit = 120

[tool.pylint]
max-line-length = 120
disable = [
    "C0114", # Missing module docstring
    "C0115", # Missing class docstring
    "C0116", # Missing function or method docstring
    "C0415", # Import outside toplevel
]
good-names = ["i", "f", "x", "y"]

[tool.pylint.typecheck]
generated-members = ["torch.*", "numpy.*", "cv2.*"]

[tool.setuptools]
packages = [
    "sign_language_segmentation",
    "sign_language_segmentation.src",
    "sign_language_segmentation.src.utils",
]

[tool.setuptools.package-data]
sign_language_segmentation = ["**/*.pth"]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["sign_language_segmentation"]

[project.scripts]
pose_to_segments = "sign_language_segmentation.bin:main"